static/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Groq API (replaces local Ollama)
groq==0.4.1

# Vector Database & Embeddings (using Groq)
faiss-cpu==1.8.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Audio Processing (keeping for now, will use Groq when available)
pydub==0.25.1
gTTS==2.5.3
pyttsx3==2.90
SpeechRecognition==3.10.0
soundfile==0.12.1
librosa==0.10.1

# NLP & Text Processing (reduced dependencies)
nltk==3.8.1
spacy==3.7.2
PyPDF2==3.0.1
pdfplumber==0.10.3

# Math & Scientific Computing
numpy==1.26.4
scipy==1.11.4
sympy==1.12
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.0
packaging==23.2
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
redis==5.0.1
celery==5.3.4

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.0
flake8==6.1.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk[flask]==1.39.1

# API Documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# File handling
python-multipart==0.0.6
werkzeug==3.0.1
